added code for multiplication of strings

+++ bignum.c
@@ -6,7 +6,7 @@
 /*   By: varnaud <varnaud@student.42.us.org>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/12/09 17:20:20 by varnaud           #+#    #+#             */
-/*   Updated: 2017/01/15 18:36:14 by varnaud          ###   ########.fr       */
+/*   Updated: 2017/01/15 23:05:51 by varnaud          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ static char	addchars(char n1, char n2, char *reminder)
 	return (r + '0');
 }
 
-char		*ft_bignum_mtp(const char *n1, int by)
+char		*ft_bignum_mtp(const char *n1, unsigned long long by)
 {
 	char	*result;
 	char	*tmp;
